FOODI - ORDER ONLINE
Welcome to Foodi-Express, the codebase for an efficient food ordering website! With a blend of HTML, CSS, and JavaScript, Foodi-Express offers a delightful user experience for exploring and ordering your favorite dishes. Whether you're a food enthusiast or a developer, this repository provides a solid foundation to build upon. Elevate your culinary journey with Foodi-Express – easy, fast, and flavorful!